combine provider serializers into single file
