+++ api/v2/serializers/__init__.py
@@ -6,14 +6,11 @@ from .instance_summary_serializer import InstanceSummarySerializer
 from .volume_serializer import VolumeSerializer
 from .volume_summary_serializer import VolumeSummarySerializer
 from .image_serializer import ImageSerializer
-from .provider_serializer import ProviderSerializer
+from .provider import ProviderSerializer, ProviderSummarySerializer, ProviderTypeSerializer, PlatformTypeSerializer
 from .identity import IdentitySerializer, IdentitySummarySerializer
 from .quota_serializer import QuotaSerializer
 from .allocation_serializer import AllocationSerializer
-from .provider_type_serializer import ProviderTypeSerializer
-from .platform_type_serializer import PlatformTypeSerializer
 from .provider_machine_serializer import ProviderMachineSerializer
 from .image_bookmark_serializer import ImageBookmarkSerializer
 from .size_serializer import SizeSerializer
-from .provider_summary_serializer import ProviderSummarySerializer
 from .size_summary_serializer import SizeSummarySerializer
